Yalebot-master/requirements.txt
Flask
gunicorn==19.6.0
zalgoify>=1.0
upsidedown
pillow
GroupyAPI
beautifulsoup4

cmake
face_recognition
scikit-image
